refactor: replace useVariantsProgressInfo hook with getVariantsProgressInfo function for variant status handling
